requirements.txt
azure-cli-core
subprocess.run
argparse
